flask-api-blog/pyproject.toml
[project]
name = "flask-blog-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Glayton",email = "gr.martins@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "apispec-webframeworks (>=1.2.0,<2.0.0)",
    "flask-swagger-ui (>=5.21.0,<6.0.0)",
    "flask-marshmallow (>=1.3.0,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "flask_blog_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
